requirements.txt
boto3>=1.36.16,<1.37.0
pandas==2.2.3
streamlit==1.43.0
awswrangler==3.11.0
loguru==0.7.3
python-dotenv>=1.0.0


aws-cdk-lib==2.177.0
aws-cdk.aws-lambda-python-alpha==2.177.0a0
constructs>=10.4.2,<11.0.0


cdklabs.generative-ai-cdk-constructs==0.1.292
aws-lambda-powertools==3.2.0
pydantic==2.9.2

# Optional:
# faker==30.8.2
# amazon-textract-textractor==1.8.5
# s3fs==2024.10.0